Fix and centralize GeoJson serialization for js (georefs) #4383
